+++ src/render/graph.rs
@@ -469,7 +469,7 @@ mod tests {
     fn config() -> ChannelConfig {
         crate::node::ChannelConfigOptions {
             count: 2,
-            mode: crate::node::ChannelCountMode::Explicit,
+            count_mode: crate::node::ChannelCountMode::Explicit,
             interpretation: crate::node::ChannelInterpretation::Speakers,
         }
         .into()